requirements.txt
openai-whisper>=20231117
google-generativeai>=0.3.0
google-cloud-speech>=2.21.0
pydantic>=2.0.0
pydub>=0.25.1
torch>=2.0.0
pyyaml>=6.0
aiofiles>=23.0
pytest>=7.0
